Extending ICE to Events (N-ary predicates)
This is a major update to ICE incorporating the code for recognizing ‘events’ (elementary clauses) headed by a verb with multiple arguments.  These are represented as first-class objects (IceTrees), in contrast to the original code for DependencyPaths, which are currently represented internally by Strings.  Only the code for collecting events from corpora and active learning are included;  code for extraction based on these clauses must still be added.

Most of the new code was added to the 'events' directory as separate classes, with the intent of later merging the processing of dependency paths and IceTrees.

In addition a modification is included (primarily to DepPaths) to speed up regeneration of LDPs and IceTrees after the entity sets have been modified.  In the modified version LDPs and IceTrees are regenerated only for documents containing words in the entity sets.